Claudwise/requirements.txt
streamlit>=1.28.0
transformers>=4.35.0
torch>=2.0.0
huggingface-hub>=0.19.0
spacy>=3.7.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
plotly>=5.17.0
streamlit-option-menu>=0.3.0
cryptography>=41.0.0
bcrypt>=4.0.0
